requirements.txt
numpy
requests
opencv-python-headless
Flask==3.0.3
flask_cors
pillow
